Move game logic into separate package
