+++ app/src/main/java/com/kamron/pogoiv/PokemonSpinnerAdapter.java
@@ -7,6 +7,8 @@ import android.view.ViewGroup;
 import android.widget.ArrayAdapter;
 import android.widget.TextView;
 
+import com.kamron.pogoiv.logic.Pokemon;
+
 import java.util.ArrayList;
 
 /**
