+++ app/src/main/java/com/kamron/pogoiv/logic/PokeInfoCalculator.java
@@ -1,4 +1,6 @@
-package com.kamron.pogoiv;
+package com.kamron.pogoiv.logic;
+
+import com.kamron.pogoiv.UpgradeCost;
 
 import java.util.ArrayList;
 import java.util.Collections;
