+++ app/src/main/java/com/kamron/pogoiv/logic/Pokemon.java
@@ -1,4 +1,4 @@
-package com.kamron.pogoiv;
+package com.kamron.pogoiv.logic;
 
 import java.util.ArrayList;
 import java.util.List;
